+++ anymail/backends/amazon_sesv1.py
@@ -1,16 +1,18 @@
-import email.charset
-import email.encoders
-import email.policy
+import warnings
+from email.charset import QP, Charset
+from email.mime.text import MIMEText
 
-from .. import __version__ as ANYMAIL_VERSION
-from ..exceptions import AnymailAPIError, AnymailImproperlyInstalled
+from ..exceptions import (
+    AnymailAPIError,
+    AnymailDeprecationWarning,
+    AnymailImproperlyInstalled,
+)
 from ..message import AnymailRecipientStatus
 from ..utils import UNSET, get_anymail_setting
 from .base import AnymailBaseBackend, BasePayload
 
 try:
     import boto3
-    from botocore.client import Config
     from botocore.exceptions import BotoCoreError, ClientError, ConnectionError
 except ImportError as err:
     raise AnymailImproperlyInstalled(
@@ -24,18 +26,27 @@ BOTO_BASE_ERRORS = (BotoCoreError, ClientError, ConnectionError)
 
 class EmailBackend(AnymailBaseBackend):
     """
-    Amazon SES v2 Email Backend (using boto3)
+    Amazon SES v1 Email Backend (using boto3)
     """
 
     esp_name = "Amazon SES"
 
     def __init__(self, **kwargs):
         """Init options from Django settings"""
+        from .amazon_ses import _get_anymail_boto3_params
+
+        warnings.warn(
+            "anymail.backends.amazon_sesv1.EmailBackend is deprecated"
+            " and will be removed in the near future. Please migrate"
+            " to anymail.backends.amazon_ses.EmailBackend using Amazon SES v2.",
+            AnymailDeprecationWarning,
+        )
         super().__init__(**kwargs)
+
         # AMAZON_SES_CLIENT_PARAMS is optional
         # (boto3 can find credentials several other ways)
         self.session_params, self.client_params = _get_anymail_boto3_params(
-            esp_name=self.esp_name, kwargs=kwargs
+            kwargs=kwargs
         )
         self.configuration_set_name = get_anymail_setting(
             "configuration_set_name",
@@ -51,7 +62,6 @@ class EmailBackend(AnymailBaseBackend):
             allow_bare=False,
             default=None,
         )
-
         self.client = None
 
     def open(self):
@@ -59,7 +69,7 @@ class EmailBackend(AnymailBaseBackend):
             return False  # already exists
         try:
             self.client = boto3.session.Session(**self.session_params).client(
-                "sesv2", **self.client_params
+                "ses", **self.client_params
             )
         except Exception:
             if not self.fail_silently:
@@ -70,7 +80,7 @@ class EmailBackend(AnymailBaseBackend):
     def close(self):
         if self.client is None:
             return
-        self.client.close()
+        # self.client.close()  # boto3 doesn't support (or require) client shutdown
         self.client = None
 
     def _send(self, message):
@@ -90,24 +100,16 @@ class EmailBackend(AnymailBaseBackend):
             )
 
     def build_message_payload(self, message, defaults):
+        # The SES SendRawEmail and SendBulkTemplatedEmail calls have
+        # very different signatures, so use a custom payload for each
         if getattr(message, "template_id", UNSET) is not UNSET:
-            # For simplicity, use SESv2 SendBulkEmail for all templated messages
-            # (even though SESv2 SendEmail has a template option).
-            return AmazonSESV2SendBulkEmailPayload(message, defaults, self)
+            return AmazonSESSendBulkTemplatedEmailPayload(message, defaults, self)
         else:
-            return AmazonSESV2SendEmailPayload(message, defaults, self)
+            return AmazonSESSendRawEmailPayload(message, defaults, self)
 
     def post_to_esp(self, payload, message):
-        payload.finalize_payload()
         try:
-            client_send_api = getattr(self.client, payload.api_name)
-        except AttributeError:
-            raise NotImplementedError(
-                f"boto3 sesv2 client does not have method {payload.api_name!r}."
-                f" Check {payload.__class__.__name__}.api_name."
-            ) from None
-        try:
-            response = client_send_api(**payload.params)
+            response = payload.call_send_api(self.client)
         except BOTO_BASE_ERRORS as err:
             # ClientError has a response attr with parsed json error response
             # (other errors don't)
@@ -125,70 +127,64 @@ class EmailBackend(AnymailBaseBackend):
 
 
 class AmazonSESBasePayload(BasePayload):
-    #: Name of the boto3 SES/SESv2 client method to call
-    api_name = "SUBCLASS_MUST_OVERRIDE"
-
     def init_payload(self):
         self.params = {}
         if self.backend.configuration_set_name is not None:
             self.params["ConfigurationSetName"] = self.backend.configuration_set_name
 
-    def finalize_payload(self):
-        pass
+    def call_send_api(self, ses_client):
+        raise NotImplementedError()
 
     def parse_recipient_status(self, response):
         # response is the parsed (dict) JSON returned from the API call
         raise NotImplementedError()
 
     def set_esp_extra(self, extra):
-        # e.g., ConfigurationSetName, FromEmailAddressIdentityArn,
-        # FeedbackForwardingEmailAddress, ListManagementOptions
+        # e.g., ConfigurationSetName, FromArn, SourceArn, ReturnPathArn
         self.params.update(extra)
 
 
-class AmazonSESV2SendEmailPayload(AmazonSESBasePayload):
-    api_name = "send_email"
-
+class AmazonSESSendRawEmailPayload(AmazonSESBasePayload):
     def init_payload(self):
         super().init_payload()
-        self.all_recipients = []  # for parse_recipient_status
+        self.all_recipients = []
         self.mime_message = self.message.message()
 
-    def finalize_payload(self):
-        # (The boto3 SES client handles base64 encoding raw_message.)
-        raw_message = self.generate_raw_message()
-        self.params["Content"] = {"Raw": {"Data": raw_message}}
-
-    def generate_raw_message(self):
-        """
-        Serialize self.mime_message as an RFC-5322/-2045 MIME message,
-        encoded as 7bit-clean, us-ascii byte data.
-        """
-        # Amazon SES does not support `Content-Transfer-Encoding: 8bit`. And using 8bit
-        # with SES open or click tracking results in mis-encoded characters. To avoid
-        # this, convert any 8bit parts to 7bit quoted printable or base64. (We own
-        # self.mime_message, so destructively modifying it should be OK.)
-        # (You might think cte_type="7bit" in the email.policy below would cover this,
-        # but it seems that cte_type is only examined as the MIME parts are constructed,
-        # not when an email.generator serializes them.)
+        # Work around an Amazon SES bug where, if all of:
+        #   - the message body (text or html) contains non-ASCII characters
+        #   - the body is sent with `Content-Transfer-Encoding: 8bit`
+        #     (which is Django email's default for most non-ASCII bodies)
+        #   - you are using an SES ConfigurationSet with open or click tracking enabled
+        # then SES replaces the non-ASCII characters with question marks as it rewrites
+        # the message to add tracking. Forcing `CTE: quoted-printable` avoids the
+        # problem. (https://forums.aws.amazon.com/thread.jspa?threadID=287048)
         for part in self.mime_message.walk():
-            if part["Content-Transfer-Encoding"] == "8bit":
+            if (
+                part.get_content_maintype() == "text"
+                and part["Content-Transfer-Encoding"] == "8bit"
+            ):
+                content = part.get_payload()
                 del part["Content-Transfer-Encoding"]
-                if part.get_content_maintype() == "text":
-                    # (Avoid base64 for text parts, which can trigger spam filters)
-                    email.encoders.encode_quopri(part)
-                else:
-                    email.encoders.encode_base64(part)
+                qp_charset = Charset(part.get_content_charset("us-ascii"))
+                qp_charset.body_encoding = QP
+                # (can't use part.set_payload, because SafeMIMEText can undo
+                # this workaround)
+                MIMEText.set_payload(part, content, charset=qp_charset)
 
-        self.mime_message.policy = email.policy.default.clone(cte_type="7bit")
-        return self.mime_message.as_bytes()
+    def call_send_api(self, ses_client):
+        # Set Destinations to make sure we pick up all recipients (including bcc).
+        # Any non-ASCII characters in recipient domains must be encoded with Punycode.
+        # (Amazon SES doesn't support non-ASCII recipient usernames.)
+        self.params["Destinations"] = [email.address for email in self.all_recipients]
+        self.params["RawMessage"] = {"Data": self.mime_message.as_bytes()}
+        return ses_client.send_raw_email(**self.params)
 
     def parse_recipient_status(self, response):
         try:
             message_id = response["MessageId"]
         except (KeyError, TypeError) as err:
             raise AnymailAPIError(
-                f"{err!s} parsing Amazon SES send result {response!r}",
+                "%s parsing Amazon SES send result %r" % (str(err), response),
                 backend=self.backend,
                 email_message=self.message,
                 payload=self,
@@ -210,31 +206,22 @@ class AmazonSESV2SendEmailPayload(AmazonSESBasePayload):
         # merged/computed value isn't forced back into the EmailMessage.
         if attr in self.defaults:
             self.unsupported_feature(
-                f"Anymail send defaults for '{attr}' with Amazon SES"
+                "Anymail send defaults for '%s' with Amazon SES" % attr
             )
 
-    def set_from_email(self, email):
-        # If params["FromEmailAddress"] is not provided, SES will parse it from the raw
-        # mime_message headers. (And setting it replaces any From header. Note that
-        # v2 SendEmail doesn't have an equivalent to v1 SendRawEmail's Sender param.)
-        self._no_send_defaults("from_email")
+    def set_from_email_list(self, emails):
+        # Although Amazon SES will send messages with any From header, it can only parse
+        # Source if the From header is a single email. Explicit Source avoids an
+        # "Illegal address" error:
+        if len(emails) > 1:
+            self.params["Source"] = emails[0].addr_spec
+        # (else SES will look at the (single) address in the From header)
 
     def set_recipients(self, recipient_type, emails):
-        # Although Amazon SES can parse the 'to' and 'cc' recipients from the raw
-        # mime_message headers, providing them in the Destination param makes it
-        # explicit (and is required for 'bcc' and for spoofed 'to').
-        self.all_recipients += emails  # save for parse_recipient_status
+        self.all_recipients += emails
+        # included in mime_message
+        assert recipient_type in ("to", "cc", "bcc")
         self._no_send_defaults(recipient_type)
-
-        if emails:
-            # params["Destination"] = {"ToAddresses": [...], "CcAddresses": etc.}
-            # (Unlike most SendEmail params, these _don't_ replace the corresponding
-            # raw mime_message headers.)
-            assert recipient_type in ("to", "cc", "bcc")
-            destination_key = f"{recipient_type.capitalize()}Addresses"
-            self.params.setdefault("Destination", {})[destination_key] = [
-                email.address for email in emails
-            ]
 
     def set_subject(self, subject):
         # included in mime_message
@@ -242,7 +229,6 @@ class AmazonSESV2SendEmailPayload(AmazonSESBasePayload):
 
     def set_reply_to(self, emails):
         # included in mime_message
-        # (and setting params["ReplyToAddresses"] replaces any Reply-To header)
         self._no_send_defaults("reply_to")
 
     def set_extra_headers(self, headers):
@@ -266,18 +252,16 @@ class AmazonSESV2SendEmailPayload(AmazonSESBasePayload):
         self._no_send_defaults("attachments")
 
     # Anymail-specific payload construction
-
     def set_envelope_sender(self, email):
-        # Amazon SES will generate a unique mailfrom, and then forward any delivery
-        # problem reports that address receives to the address specified here:
-        self.params["FeedbackForwardingEmailAddress"] = email.addr_spec
+        self.params["Source"] = email.addr_spec
 
     def set_spoofed_to_header(self, header_to):
         # django.core.mail.EmailMessage.message() has already set
         #   self.mime_message["To"] = header_to
         # and performed any necessary header sanitization.
         #
-        # The actual "to" is already in params["Destination"]["ToAddresses"].
+        # The actual "to" is already in self.all_recipients,
+        # which is used as the SendRawEmail Destinations later.
         #
         # So, nothing to do here, except prevent the default
         # "unsupported feature" error.
@@ -316,13 +300,13 @@ class AmazonSESV2SendEmailPayload(AmazonSESBasePayload):
                 self.unsupported_feature(
                     "multiple tags with the AMAZON_SES_MESSAGE_TAG_NAME setting"
                 )
-            self.params.setdefault("EmailTags", []).append(
+            self.params.setdefault("Tags", []).append(
                 {"Name": self.backend.message_tag_name, "Value": tags[0]}
             )
 
     def set_template_id(self, template_id):
         raise NotImplementedError(
-            f"{self.__class__.__name__} should not have been used with template_id"
+            "AmazonSESSendRawEmailPayload should not have been used with template_id"
         )
 
     def set_merge_data(self, merge_data):
@@ -332,19 +316,15 @@ class AmazonSESV2SendEmailPayload(AmazonSESBasePayload):
         self.unsupported_feature("global_merge_data without template_id")
 
 
-class AmazonSESV2SendBulkEmailPayload(AmazonSESBasePayload):
-    api_name = "send_bulk_email"
-
+class AmazonSESSendBulkTemplatedEmailPayload(AmazonSESBasePayload):
     def init_payload(self):
         super().init_payload()
-        # late-bind recipients and merge_data in finalize_payload
+        # late-bind recipients and merge_data in call_send_api
         self.recipients = {"to": [], "cc": [], "bcc": []}
         self.merge_data = {}
 
-    def finalize_payload(self):
-        # Build BulkEmailEntries from recipients and merge_data.
-
-        # Any cc and bcc recipients should be included in every entry:
+    def call_send_api(self, ses_client):
+        # include any 'cc' or 'bcc' in every destination
         cc_and_bcc_addresses = {}
         if self.recipients["cc"]:
             cc_and_bcc_addresses["CcAddresses"] = [
@@ -355,75 +335,55 @@ class AmazonSESV2SendBulkEmailPayload(AmazonSESBasePayload):
                 bcc.address for bcc in self.recipients["bcc"]
             ]
 
-        # Construct an entry with merge data for each "to" recipient:
-        self.params["BulkEmailEntries"] = [
+        # set up destination and data for each 'to'
+        self.params["Destinations"] = [
             {
                 "Destination": dict(ToAddresses=[to.address], **cc_and_bcc_addresses),
-                "ReplacementEmailContent": {
-                    "ReplacementTemplate": {
-                        "ReplacementTemplateData": self.serialize_json(
-                            self.merge_data.get(to.addr_spec, {})
-                        ),
-                    }
-                },
+                "ReplacementTemplateData": self.serialize_json(
+                    self.merge_data.get(to.addr_spec, {})
+                ),
             }
             for to in self.recipients["to"]
         ]
 
+        return ses_client.send_bulk_templated_email(**self.params)
+
     def parse_recipient_status(self, response):
         try:
-            results = response["BulkEmailEntryResults"]
-            ses_status_set = set(result["Status"] for result in results)
+            # response["Status"] should be a list in Destinations (to) order
             anymail_statuses = [
                 AnymailRecipientStatus(
-                    message_id=result.get("MessageId", None),
-                    status="queued" if result["Status"] == "SUCCESS" else "failed",
+                    message_id=status.get("MessageId", None),
+                    status="queued" if status.get("Status") == "Success" else "failed",
                 )
-                for result in results
+                for status in response["Status"]
             ]
         except (KeyError, TypeError) as err:
             raise AnymailAPIError(
-                f"{err!s} parsing Amazon SES send result {response!r}",
+                "%s parsing Amazon SES send result %r" % (str(err), response),
                 backend=self.backend,
                 email_message=self.message,
                 payload=self,
             ) from None
 
-        # If all BulkEmailEntryResults[].Status are the same non-success status,
-        # raise an APIError to expose the error message/reason (matching behavior
-        # of non-template SendEmail call).
-        if len(ses_status_set) == 1 and ses_status_set != {"SUCCESS"}:
-            raise AnymailAPIError(
-                # use Error text if available, else the Status enum, from first result
-                results[0].get("Error", results[0]["Status"]),
-                backend=self.backend,
-                email_message=self.message,
-                payload=self,
-                response=response,
-            )
-
-        # Otherwise, return per-recipient status (just "queued" or "failed") for
-        # all-success, mixed success/error, or all-error mixed-reason cases.
-        # The BulkEmailEntryResults are in the same order as the Destination param
-        # (which is in the same order as recipients["to"]).
         to_addrs = [to.addr_spec for to in self.recipients["to"]]
         if len(anymail_statuses) != len(to_addrs):
             raise AnymailAPIError(
-                f"Sent to {len(to_addrs)} destinations,"
-                f" but only {len(anymail_statuses)} statuses"
-                f" in Amazon SES send result {response!r}",
+                "Sent to %d destinations, but only %d statuses in Amazon SES"
+                " send result %r" % (len(to_addrs), len(anymail_statuses), response),
                 backend=self.backend,
                 email_message=self.message,
                 payload=self,
             )
+
         return dict(zip(to_addrs, anymail_statuses))
 
     def set_from_email(self, email):
         # this will RFC2047-encode display_name if needed:
-        self.params["FromEmailAddress"] = email.address
+        self.params["Source"] = email.address
 
     def set_recipients(self, recipient_type, emails):
-        # late-bound in finalize_payload
+        # late-bound in call_send_api
         assert recipient_type in ("to", "cc", "bcc")
         self.recipients[recipient_type] = emails
 
@@ -452,27 +412,24 @@ class AmazonSESV2SendBulkEmailPayload(AmazonSESBasePayload):
             self.unsupported_feature("attachments with template")
 
     # Anymail-specific payload construction
-
     def set_envelope_sender(self, email):
-        # Amazon SES will generate a unique mailfrom, and then forward any delivery
-        # problem reports that address receives to the address specified here:
-        self.params["FeedbackForwardingEmailAddress"] = email.addr_spec
+        self.params["ReturnPath"] = email.addr_spec
 
     def set_metadata(self, metadata):
-        # no custom headers with SendBulkEmail
+        # no custom headers with SendBulkTemplatedEmail
         self.unsupported_feature("metadata with template")
 
     def set_tags(self, tags):
-        # no custom headers with SendBulkEmail, but support
+        # no custom headers with SendBulkTemplatedEmail, but support
         # AMAZON_SES_MESSAGE_TAG_NAME if used (see tags/metadata in
-        # AmazonSESV2SendEmailPayload for more info)
+        # AmazonSESSendRawEmailPayload for more info)
         if tags:
             if self.backend.message_tag_name is not None:
                 if len(tags) > 1:
                     self.unsupported_feature(
                         "multiple tags with the AMAZON_SES_MESSAGE_TAG_NAME setting"
                     )
-                self.params["DefaultEmailTags"] = [
+                self.params["DefaultTags"] = [
                     {"Name": self.backend.message_tag_name, "Value": tags[0]}
                 ]
             else:
@@ -482,53 +439,11 @@ class AmazonSESV2SendBulkEmailPayload(AmazonSESBasePayload):
                 )
 
     def set_template_id(self, template_id):
-        # DefaultContent.Template.TemplateName
-        self.params.setdefault("DefaultContent", {}).setdefault("Template", {})[
-            "TemplateName"
-        ] = template_id
+        self.params["Template"] = template_id
 
     def set_merge_data(self, merge_data):
-        # late-bound in finalize_payload
+        # late-bound in call_send_api
         self.merge_data = merge_data
 
     def set_merge_global_data(self, merge_global_data):
-        # DefaultContent.Template.TemplateData
-        self.params.setdefault("DefaultContent", {}).setdefault("Template", {})[
-            "TemplateData"
-        ] = self.serialize_json(merge_global_data)
-
-
-def _get_anymail_boto3_params(esp_name=EmailBackend.esp_name, kwargs=None):
-    """Returns 2 dicts of params for boto3.session.Session() and .client()
-
-    Incorporates ANYMAIL["AMAZON_SES_SESSION_PARAMS"] and
-    ANYMAIL["AMAZON_SES_CLIENT_PARAMS"] settings.
-
-    Converts config dict to botocore.client.Config if needed
-
-    May remove keys from kwargs, but won't modify original settings
-    """
-    # (shared with ..webhooks.amazon_ses)
-    session_params = get_anymail_setting(
-        "session_params", esp_name=esp_name, kwargs=kwargs, default={}
-    )
-    client_params = get_anymail_setting(
-        "client_params", esp_name=esp_name, kwargs=kwargs, default={}
-    )
-
-    # Add Anymail user-agent, and convert config dict to botocore.client.Config
-    client_params = client_params.copy()  # don't modify source
-    config = Config(
-        user_agent_extra="django-anymail/{version}-{esp}".format(
-            esp=esp_name.lower().replace(" ", "-"), version=ANYMAIL_VERSION
-        )
-    )
-    if "config" in client_params:
-        # convert config dict to botocore.client.Config if needed
-        client_params_config = client_params["config"]
-        if not isinstance(client_params_config, Config):
-            client_params_config = Config(**client_params_config)
-        config = config.merge(client_params_config)
-    client_params["config"] = config
-
-    return session_params, client_params
+        self.params["DefaultTemplateData"] = self.serialize_json(merge_global_data)
